requirements.txt
pyspark==4.0.0
py4j==0.10.9.9
apache-airflow==2.9.2
clickhouse-connect==0.8.0
airflow-providers-clickhouse==0.0.1
clickhouse-driver==0.2.9
airflow-clickhouse-plugin